Rename tools/test_setup.py to OmeroPy/src/omero_setup.py
So that downstream repositories can make use of the PyTest
class defined here, the module will need to be packaged as
part of lib/python.
