+++ components/tools/OmeroFS/setup.py
@@ -14,8 +14,8 @@ for tools in glob.glob("../../../lib/repository/setuptools*.egg"):
     if tools.find(".".join(map(str, sys.version_info[0:2]))) > 0:
         sys.path.insert(0, os.path.abspath(tools))
 
-sys.path.append("..")
-from test_setup import PyTest
+sys.path.insert(0, "../OmeroPy/src")
+from omero_setup import PyTest
 
 from ez_setup import use_setuptools
 use_setuptools(to_dir='../../../lib/repository')
