+++ components/tools/OmeroPy/src/omero_setup.py
